Abstracted all printing behaviour to a TopicContainer class
